chore(Extension): allow TryGetComponent to get any valid type
The TryGetComponent method on GameObjectExtensions has been updated to
accept a generic type parameter so it can be used to look up any
component type on a GameObject and isn't just limited to the fixed
type of `Component`. The `FindRigidbody` extension has also been
removed and the generic logic of searching ancestors and descendants
has been moved into the `TryGetComponent` method as this single
method is capable of doing all of the required functions.