chore(conventions): apply field attributes inline
The convention for using field attributes is to write them inline
with the `Header` attribute also being a part of the first inline
attribute to ensure the code attributes conform with MSDN conventions.

All of the relevant files have been updated to apply the new
conventions, which have been described in the CONTRIBUTING.md document.