feat(Interactions): list grabbed interactables for the interactor
The InteractorFacade now has a mechanism for retrieving the current
grabbed interactable GameObjects. It achieves this by utilising a
GameObjectSet on the Interactor which is updated by the
InteractableFacade when it emits the relevant events on the Interactor
as this is the only callback from an Interactable to the relevant
Interactor.