fix(Interactions): allow nested interactables
Previously, the Interactables couldn't be nested because the
ActiveCollisionPublisher would always publish to every
ActiveCollisionConsumer that was found in the top level Interactable
meaning that all events on the top level Interactable would be pushed
to every child Interactable as well.

The fix is to allow the Consumer to specify what it's container
GameObject is and have the Publisher only publish to those that
match the containing rigidbody as the rigidbody should stop any
higher up collisions from being notified.