fix(Modification): account for pivot rotation when reset orientation
The DirectionModifier was not taking into account if the pivot rotated
when the reset orientation was performed so if the pivot rotation had
changed then the saved initial rotation was incorrect and would not
rotate correctly back to the initial position.

This fix utilises the pivot rotation values to calculate the actual
initial position to ensure the rotation reset works correctly.