fix(Modification): ensure references are not null before acting on them
The DirectionModifier could throw errors in certain edge cases as the
use of the GameObject references for target and pivot were not always
checked for null status before being used.

This fix ensures they are always prior checked and the checks have been
moved to more appropriate places to ensure cleaner code.