feat(Follow): add rigidbody rotation by force follow modifier property
A new follow modifier property dealing with rotation has been added
that applies a force at a given position to a rigidbody which creates
a torque providing a rotation to the object.

This is useful for controlling objects that are attached on a hinge
joint.

As this is using actual forces it cannot be tested via the editor
tests.