feat(Event): extract GameObject from Component and emit event
The ComponentGameObjectEmitter will extract the GameObject for a given
source component and emit the result as an event.

This new component has many shared similarities with the
NotifierContainerExtractor and PublisherContainerExtractor as they both
extract a GameObject and emit an event. So all of these classes now
extend a common BaseGameObjectEmitter which deals with holding the
state of the GameObject that has been extracted and provides the common
Extracted event that they all use.

The child classes now simply inherit from the base and override the
Extract method where necessary and set the extracted result and call
the base Extract method.