refactor(Event): add common base for restrictable event proxy emitters
A number of the EventProxyEmitters employed their own RuleContainer
implementation which all basically provided the same functionality of
checking to see if the Receive was valid.

A new abstract base class of RestrictableSingleEventProxyEmitter has
been added that wraps up that common functionality and all the relevant
event proxy emitters that require it now extend that class.