feat(Cast): add fixed line caster
The FixedLineCast extends the StraightLineCast as it casts a straight
line but forgoes the use of a ray to determine the line. It simply
casts from the origin along the forward direction to the given length.

A few changes to the PointsCast have been done to make extending
cleaner and require less duplication of code. It also means the code
doing the casting that the inherited classes use is now protected.