fix(Transformation): use vector multiplier
The Vector2Multiplier was using two floats as the multiplier to use
in the operation this type implements. A Vector2 is a better
replacement for those floats and supports the same operations. The
existing setter methods are kept so per-axis changes are possible
via Editor inspectors while this change additionally allows
specifying a Vector2 directly.