feat(Collection): abort a stack pop during process
The GameObjectEventStack can now have the pop process aborted during
the stack being popped.

For example if a stack contains 1,2,3 and 2 is popped from the stack
then at the point 3 is being force popped it can forcibly abort the
remaining popping of the stack so 1 and 2 stay on the stack and only
3 is popped.

This is useful if the stack is to be re-built during a pop process as
the original stack should therefore no longer be manipulated as there
will be a new stack generated.