feat(BodyRepresentation): CharacterController as body representation
The Unity `CharacterController` was made for non-VR usage but is
equally helpful when implementing the concept of "body physics" for
VR. A new component has been added that allows setting the height of
the controller to match a given Game Object and most importantly
allows moving the controller by observing movement to a rigidbody
that matches the controller's collider. With this component a body
representation in form of the controller is able to follow and match
a VR headset moving in the real world while adhering to the virtual
world's constraints imposed by both the physical world as well as
other components affecting the controller. A new prefab has been
added that encapsulates and sets up a body representation using this
new component.