feat(Climb): provide a prefab to allow for climbing
A new prefab and facade has been added that allows climbing by
observing and applying movement from an interactor and an
interactable. Besides offering methods to start and stop a climb and
the respective events the facade also offers an optional stop method
that applies a velocity. This allows flinging the target object
based on the incoming velocity.