feat(Climb): add climbable interactable prefab
A new prefab and facade have been added to allow an interactable to
be climbable by grabbing it. The components also take care of
applying the interactor's velocity when the climb ends to allow
"flinging". A setting on the facade allows setting the multiplier to
apply to the incoming velocity of the interactor as 1:1 movement
often doesn't feel right in VR. This setting has the added benefit
that a vector of zero disables the velocity based flinging.