feat(Locomotion): ability to move object based on given axis data
The AxisToVector3 prefab allows an object (e.g. PlayArea) to be moved
by passing on vector3 data to update the position of the Transform
by taking input from 2 axes that control the lateral
(left/right direction) and longitudinal (forward/backward direction).

This provides a simple way of moving the PlayArea around with a
controller thumbstick/touchpad axis. It's also possible to provide
a forward source GameObject to use as the direction to move the target.