feat(Collection): add observable list for game objects
A new type has been added that forms the basis for observable lists.
An observable list is a list that is a Unity component and thus
shows up in the Editor. The various events on it allow reacting to
elements getting added to or removed from the collection. One
subclass of the base type has been added to allow observing a list
of GameObjects.