feat(BodyRepresentation): allow preventing colliding movement
If the body moves in a way that doesn't allow the physics engine to
resolve any potential collision (i.e. setting the transform's
position instead of using force/velocity of the rigidbody) the body
may get stuck in something. While the physics engine will solve the
collision during the next physics loop the outcome is often unwanted
as the user may notice a "jump" in position. A method has been added
to the BodyRepresentationFacade that allows solving these body
collisions early. If the method is called in the same frame after
the movement has been applied the movement will not result in a
collision. Instead the body stops in front of walls etc. without a
single frame ever getting rendered to the user, preventing the
aforementioned "jump" in position.