feat(BodyRepresentation): ignore interactors and grabbed interactables
The BodyRepresentation script now utilises an internal ObservableList
that stores the GameObjects that colliders should be ignored with
the CharacterCollider and Rigidbody collider. This list can be
populated manually at runtime or a new helper List on the
BodyRepresentationFacade can be used.

The helper List takes a collection of InteractorFacades and for each
one ignores collisions on any child collider of the Interactor and
it also sets up a listener on the Grabbed/Ungrabbed event of the
Interactor to ensure any grabbed interactable is also ignored from
collisions with the Body Representation colliders.