fix(BodyRepresentation): upwards movement because of other colliders
When the body representation is standing or falling on top of other
colliders that are moving it often got stuck in the air or even fell
through the other collider. With this fix applied the
CharacterController is always moved, even a tiny bit in case it
doesn't actually need any movement as that triggers the
CharacterController's internal collision resolution.