feat(Interactions): perform first/last touch/grab in script
The first/last-Touch/Grab methods used to be public and would be
called from the prefab events. However, this meant that the prefab
would need to always be set up to call these intrinsic events whereas
the touch/untouch grab/ungrab methods can always know when to call
the first/last events and always know to do it in the right order
therefore avoiding any prefab configuration issues.

The Climbable prefab was not recognising grabbed interactors because
the script relied upon there being a GameObjectStack but the Climbable
interactor didn't utilize one. This has now been updated so the
Climbable uses the GameObjectSet and the internal grab setup script
can choose between either using a provided GameObjectStack or if that
is not provided then the provided GameObjectSet.

The GameObjectSet is useful when the order of interactors being added
to the interactable is not important, such as when climbing occurs.