fix(Structure): remove erroneous uses of monadic null checks
Monadic null checking was introduced in C#6 and means that objects
don't need an `if (x == null)` check and can simply do `x?.` to
determine if the object is null. However, as Unity bypasses the
standard operators and checks, this means the monadic null check
does not work with Unity MonoBehaviour scripts along with GameObject
and Transform objects so any use of `?.` will not produce the
expected results.

This fix removes any misuse of the monadic null check.