fix(Interactions): allow auto grab to work with precognition
Due to the changes introduced by the Grab Precognition it meant that
the CountdownTimer would be responsible for publishing the grab
action and this would mess up the procedure of the auto grab function.

The solution is to ensure the CountdownTimer is disabled and to
manually publish the grab action and then re-enable the CountdownTimer
and for this to work, the CountdownTimer also needs to only be able to
be started if the component is active and enabled.