feat(Velocity): calculate with optional relative to object
The VelocityEstimator now allows for a relativeTo GameObject that
can be used to determine a relative velocity rather than a global
velocity.

This is useful if the tracked object is also being moved due to
a parent object being moved.

The CameraRig prefabs have been updated so the controller velocity
is always tracked relative to the parent play area.