feat(Process): add utilization to MomentProcess
The abstraction of repeatable units of work into `IProcessable`
allows control over _when_ to actually run them without forcing
script changes. Expanding on the same abstraction a concept of
"utilization" has been introduced: Utilization of a process can now
be adjusted to control _how often_ to run the unit of work. As a
result of enabling this feature the `MomentProcess` itself is now
`IProcessable`, too, which enables chains of processing. Each part
of the chain comes with the control mechanisms enabled by
`MomentProcess`.