feat(Locomotion): allow teleport offset to be optional
The Teleporter would previously always apply an offset of the
HeadsetAlias to the PlayAreaAlias when moving it but there can be
instances where the PlayAreaAlias centre is wanted to be moved to
the teleport point.

This is now possible by optionally setting the Offset parameter
on the TransformPropertyApplier inside the HeightAdjustTeleporter
GameObject.

The parameter fields have also changed to be more generic.

PlayAreaAlias becomes Target
HeadsetAlias becomes Offset

And a new bool, of `OnlyOffsetFloorSnap` allows for the offset
on teleport to be optional. If the setting is true then the Offset
is only applied when doing floor snapping, which has to always be
applied and cannot be optional.