feat(Transformation): aggregator to subtract a Vector3 collection
The Vector3Subtractor extends the newly created base class
CollectionAggregator which allows for performing aggregate functions
on a typed collection.

Previously, there was a FloatCollectionAggregator that added up float
values, but this has now become the FloatAdder and also extends the
CollectionAggregator as the aggregation of the collections is generic
up until the required aggregate function.