feat(PlayAreaRepresentation): option to move target based on offset
The PlayAreaRepresentation is now a nested Prefab with its own Facade
to control the relevant set up. It also has the ability to set an
offset origin and offset destination to determine the offset position
to apply to the PlayAreaRepresentation target. This is useful if
the representation target is being used in the pointer cursor and it
is appropriate to show where the cursor is in relation to the playarea.