refactor(Transformation): convert multipliers to be aggregators
The various Multipler transformations now extend CollectionAggregator
as multiplication is an aggregate function across a collection of
values.

Now it's possible to multiple a range of float/vector2/vector3 values
as they extend the base aggregator.

The prefabs have been updated and the general rule is the dynamic
multiplication value is always index 0 and the multipler is always
set to index 1. This way the event updating the dynamic value never
has to bother setting the current index in the collection.