feat(Collision): add active collision validity rule
The ActiveCollisionContainer can now have a rule applied to it to
determine which collisions can be added to the collection.

This is useful for separating out collisions from a CollisionTracker
for purposes such as near touching objects where the engulfing
collider can prevent the inner collider from being notified of
touch events.