feat(Collection): add observable counter for game objects
The ObservableCounter and the concrete GameObjectObservableCounter
observe when elements are added or removed from a dictionary and
increase the count for every time the same element is added and
decrease the count for every time the same element is removed.

When a new element is added for the first time it emits an event
and when an element is removed by its counter reaching zero then
another event is emitted.

This will be used in fixing the touching/untouching state of
interactables.