fix(Interactions): touch/untouch with multi collider interactables
The touch and untouch state of a multi collider interactable has been
fixed from the previous broken state where passing the interactor
through multiple compound colliders on the same interactable would
call untouch every time the interactor left one of the compound
colliders even though the interactor was still actively touching
the interactable.

This has been fixed now by using the GameObjectObservableCounter
to keep track of the number of compound colliders within the
interactable that the interactor is touching.