feat(Process): reimplement source target processing and object follow
The SourceTargetProcessor has been re-written to be more generic and to
allow the ObjectFollower to be re-written so it is more cohesive.

The ObjectFollower now is consistent in having a collection of source
GameObjects that always are used for the source of data and a
collection of target GameObjects that always have the source data
applied to.

The source and target data can also accept a rule to determine which
source elements to use/ignore and which target elements to use/ignore
which makes it possible to replicate the original functionality of
the ObjectFollower, which was:

 * Targets follow source
 * Source follows active Target

But without the need to flip the meaning of the fields or require
enums with descriptions of confusing functionality.

Due to the changes to the SourceTargetProcessor, the
GameObjectStateMirror component also needed updating to bring it in
line with the changes.

A new ICollectionExtensions component has been added that provides
a simple helper for components that extend ICollection to retrieve
an index that is within the bounds of the collection and allows for
negative indices to make it possible to get elements starting from
the end of the collection.

The Follow modifiers have also been updated to no longer attempt
to set transform properties in local space as it makes no sense
when applying a follow operation. The offset must also be a child
of the target and always uses the offset local space in relation
to the target to offset against the source.

All of the prefabs that used a SourceTargetProcessor have been
updated to utilise the new mechanism.