feat(Collision): choose how to find collision point parent
The CollisionPointContainer would previously always use the collider's
containing GameObject as the collision point parent. However, this is
not always desired as sometimes the GameObject that contains the
CollisionNotifier component is the desired parent.

A new option has been added that allows whether to use the
CollisionNotifier GameObject as the parent.