feat(Interactions): add direction drive powered by a physics joint
The DirectionalJointDrive creates a linear motion drive via a physics
joint to move a control in a limited linear axis backed by a
configurable joint so it works within the physics system of the engine.

The Drive concept has been abstracted as other drive types that don't
utilize physics joints can be created and provide the same outcome.