feat(MovementAmplifier): amplify movement of a game object
The movement of a game object can be amplified by applying the same
movement to another object. The feature comes with a multiplier
allows setting the amplification amount as well as a threshold that
allows ignoring movement in a radius. In VR experiences this feature
can be used to amplify movement of the headset to allow traveling
larger distances than the physical play area allows. The threshold
radius prevents amplifying movement while the user is not moving and
instead just crouches.