fix(Follow): last state order in ObjectDistanceComparator
The `ObjectDistanceComparator` script remembers the state so it can
raise events when the state changes. The current implementation
remembers the state after raising the events which is an issue in
case an event listener acts on the comparator in a way that changes
the state variable, i.e. it calls
`ObjectDistanceComparator.SavePosition` or calls the setter for
`target` or `source`. The fix is to remember the state before
raising the events.