fix(Haptics): ensure active haptic process is set from list
The HapticProcessor was still in a situation where the active process
may not be active because it was setting the default one in the
OnEnable method but this was still checking to see if the process
was enabled and it could be in a situation where the process was
enabled after the HapticProcessor was enabled. This would mean the
ActiveProcess was null and the HapticProcessor did not work as expected
in these rare edge cases.

The fix is to mimic the way the VelocityTrackerProcessor handles the
active component by using a public property that does the active check.