feat(Pointer): nest common prefab functionality in object pointers
The ObjectPointer prefabs have now been updated to split out the common
functionality in the prefabs into a new nested prefab, which is
supported in Unity 3d 2018.3.

The facade has been updated so any field that affects the prefab can
be updated at runtime via a Property and will automatically
reconfigure the underlying prefab dependencies.