feat(Interactions): nest common prefab functionality in interactors
The interactor prefab has now been updated to break the internal
functionality up into appropriate nested prefabs so both the Touch
logic as well as the Grab logic now come via their own prefabs.

This makes it easier to showcase how to provide custom actions (such
as a use action) to be passed from the Interactor over to an
Interactable.

The facade has been updated so any field that affects the prefab can
be updated at runtime via a Property and will automatically
reconfigure the underlying prefab dependency.