fix(Collision): ensure container not null on destroy collision point
The CollisionPointContainer component would always run the Destroy
method even if there was no Container to destroy, meaning an empty
event would be emitted. This has been wrapped to ensure Destroy is
only called if the Container is not null.

An extra test has also been added to ensure a new container cannot
be created whilst an existing container is available.