feat(Interactions): nest common prefab functionality in interactables
The Interactable prefabs have now been split up into a more modular
design with the common logic split out into their own prefabs so it
can be reused across the different interactable types.

The action to perform on the interactable has also been split out into
their own sub prefabs so actions are easier to write and to swap in
and out of an existing interactable.

As the Interactable prefabs fundamentally changed, this also meant that
the Climb Interactable prefab needed to be updated to reflect the new
changes. The Controllable prefabs also needed updating to use the new
Interactables and now are also nested prefabs which will make
maintaining them much easier.