refactor(CameraRig): rename config component as an enactor
The previous name of the UnityXR Configuration component did not
convey that it had an active role to configure the UnityXR camera rig
so it has been renamed to make it sound more like it has an active
role therefore it is now knows as the UnityXRConfigurator i.e. the
role of the component is to Configure the UnityXR CameraRig.