refactor(CameraRig): use correct access modifiers on Simulator
The Simulator prefab has been updated so the relevant components use
the correct access modifiers for fields and provides private backing
fields for public properties to allow changing of data at runtime.

The SimulatorConfiguration component has also been renamed to
SimulatorFacade so it falls in line with other prefabs that have a
front end component that simply offers wiring up of sub components.