refactor(BodyRepresentation): move logic responsibility to internals
The BodyRepresentation prefab has been reworked so the actual logic
is inside the internal setup component rather than the facade and the
facade simply passes on public method calls to the internal script.

The fields on the facade have been updated so they can be changed at
runtime by providing public properties and private backing serialized
fields.

The internal setup component has now had all of the fields changed
to the appropriate access modifier. The `rigidbody` field has also
been renamed to `physicsBody` as using a pre-defined term can be
confusing when reading it in the codebase. Also, any external
dependency on the internal setup has now been removed and have to
go through the facade to access underlying data.