fix(structure): apply coding conventions
The updated coding conventions require parameter names to be
descriptive and easily understandable. This means that all of
the `for` loops that just create a variable called `i` have been
updated so that indexer variable is clear as to what it is doing.

The use of the variable name `@object` has also been replaced as
in many cases there is a more descriptive name that can be used
which reduces the confusion of when people are not aware of what
the preceeding `@` means in a variable name.

The only instance where the `@object` variable still exists is in
the `InterfaceContainer` where it still is the most appropriate
name for the variable.