fix(Operation): replace negative with positive connotation of axis lock
The TransformPropertyMutator had a Vector3State called `lockAxis` which
had the negative connotation that any checked box would not perform the
mutation on the axis. Other uses are positive connotations i.e. the
checked axis box means an operation **would** be performed on that
axis.

The `lockAxis` field has been changed for a `mutateOnAxis` field which
uses the positive connotation of any checked box will be mutated.

The prefabs that used the TransformPropertyMutators have also been
updated so the logic is flipped.