chore(structure): remove all prefabs
The prefabs are being removed from the core repo and moved into the
examples repo as the prefabs form the basis of examples of how to
use these design patterns and therefore are not the only use of these
design patterns. This makes it easier to recommend using this repo
for other uses other than their original intended purpose.