chore(structure): replace clearance instructions with Malimbe
Malimbe allows to automatically create simple methods that clear a
field of reference type. This change replaces existing clearance
instructions with the Malimbe attribute that allows a member to opt
into this behavior. Despite Malimbe inserting the same instruction
into the end of the method instead of exactly at the same spot the
instruction was so far this is not a breaking change as none of the
changes results in a behavioral change.