fix(Tracking): ensure correct transformation order is followed
There was an issue with the TransformPropertyApplier where it would
calculate positions incorrectly if an offset rotation was applied.

This was fundamentally down to the order or transformation being wrong
and the wrong data being passed into the position calculation.

This component has also been overhauled to make it much easier to read
and understand.