chore(structure): use serialized properties via Malimbe
Instead of writing a property and a backing field manually Malimbe
allows to just annotate the property with an attribute to ensure the
backing field is serialized. Besides that change the usages of the
updated members also have been updated and methods were added that
ensure changes to the property are handled in a boilerplate-free
way. This includes moving the property logic into "OnChange" methods
with the help of Malimbe to prevent having to write the property and
backing field combination if all that's needed is customizing the
getter or setter.