fix(Modification): add threshold to destination equality check on lerp
The TransformPropertyApplier could get into a situation where the
elapsed time would be a significantly small enough number away from the
transition duration to prevent the lerp from actually finishing and
would continue to update the properties of the transform by a very tiny
amount.

This fix adds a new threshold to check that the start and destination
properties of the transform scale, rotation and position are considered
equal to force the transition coroutine to finish and set the
properties to their final destination values.