fix(Data): only emit ObservableList events on actual change
`ObservableList` was incorrectly raising events when no actual
change happened. The fix is to only raise the events when an object
was actually found when removing and on clearing there has to be at
least one object in the list. This fix also ensures the order of the
raised events is always the same when adding: `ElementAdded` is now
fired **before** `BecamePopulated` as any other order is unexpected
to listeners.

The tests have been updated to test these changes.