feat(Data): add events to ObservableSet and allow serializing
`ObservableSet` so far was similar to `ObservableList` but didn't
allow specifying the elements at design time (because a generic set
can't be serialized in Unity), nor did it offer events that are
raised when elements are added or removed. Also, some of the
existing events on it were raised even though no actual change
happened, e.g. clearing the set without it having any elements was
firing the `BecameEmpty` event.

This change allows serializing the elements by using a private list
for this purpose (only). Additionally the missing added and removed
events have been added and the order the events are raised was
updated to be the same as in `ObservableList`. To correctly raise
events for the serialized elements a `Start` method implementation
has been added, again similar to `ObservableList`.

The tests have been updated to test these changes.