perf(general): expect collections to be non-null implicitly
By ensuring that fields of collection type are initialized with a
non-null value by default the codebase can expect non-null values to
be used from there on. Instead of handling the rare case of the
field holding the collection to be set to null it will now throw
when accessing the field instead. This performance optimization
removes the need for an extension method that would lead to
garbage-collected allocations which have to be prevented in
performance-critical contexts which can be anywhere as Zinnia is a
general-purpose library. Thus the extension method has been deleted.