perf(general): reduce heap allocations by removing LINQ usages
Heap allocations are garbage-collected which is an issue for
real-time applications. To ensure the best performance is available
to a user of Zinnia this change removes most usages of LINQ.
`foreach` loops are left in as there are no garbage allocations in
recent versions of Unity thanks to the upgraded compiler it's
targeting nowadays.