perf(general): reduce heap allocations by removing boxing of enums
`Enum.HasFlag` is a useful method to check whether a given flags
enum has the bitfield set for a specific enum variant, but it
generates a heap allocation because of the cast to `Enum`, which is
a reference-type - the enum value is boxed. This change replaces all
usages of `Enum.HasFlag` with the equivalent bitwise test, which is
allocation-free.