fix(Data): convert FloatRange to struct to prevent instancing issues
There were times where using the FloatRange.Zero helper would return
the same instance across different instantiations which won't happen
if the FloatRange is a struct instead of a class.

It makes more sense for the FloatRange to be a struct as other custom
Types included are structs unless a class is required.

Due to the FloatRange becoming a struct means a default constructor
is no longer provided which set the value to float.MinValue and
float.MaxValue so a new static helper called FloatRange.MinMax has been
added that returns the same data. Whereas a `new FloatRange()` now will
return the default values for min, max (which is 0f, 0f) so there is no
need to use the `FloatRange.Zero` helper now so that has been removed.