perf(Utility): look up interface on UnityEngine.Object only once
`InterfaceContainer` allows setting a `UnityEngine.Object` field
(which is serialized) to a type that implements a specific
interface. The check whether the set value's type implements the
interface was done every time the property getter was called. This
change ensures that lookup is only done when the field actually
changes.