perf(general): reduce heap allocations by removing boxing
By casting a value type to `object` or an interface the value will
be boxed, i.e. it wraps the value inside a `System.Object` instance
and stores it on the (managed) heap. This change reduces the amount
of heap allocations by removing a single boxing instruction still
done inside Zinnia.