perf(Association): reduce heap allocations by avoiding IEnumerable
`IEnumerable`, while exactly specifying the needed abilities
(enumeration), does not allow for a specified implementation of the
method that returns the enumerator. While classes like `List` offer
a value-typed enumerator, the default implementation for
`IEnumerable` has to create a reference type instead. This in turn
means a heap allocation which this change prevents as the changed
method always gets passed a `List`.