feat(Visual): update all fields to serialized properties
All usages of a field have now been updated to use a serialized
property to allow setting of data via UnityEvents.

New helper methods have been implemented to allow changes to
properties via code or the inspector to react to the changes and
update any relevant components as required.

The CameraColorOverlay component now also uses a working material to
apply the color overlay to instead of working directly on the injected
material. This means the injected material is used as a blueprint but
is never changed at runtime.

This new working material is automatically updated anytime the
OverlayMaterial property is changed.