feat(Action): update all fields to serialized properties
All usages of a field have now been updated to use a serialized
property to allow setting of data via UnityEvents.

New helper methods have been implemented to allow changes to
properties via code or the inspector to react to the changes and
update any relevant components as required. The compound actions
have been updated to be reactive, instead of using an `Update` loop.

Additionally the tests have been updated to verify the new logic.