feat(Utility): update all fields to be serialized properties
All usages of fields have now been switched to properties that
are serialized with Malimbe to create an appropriate backing field
to display in the Unity Inspector.

New helper methods have been implemented to allow changes to
properties via code or the inspector to react to the changes and
update any relevant components as required.

Also, now the tests run in play mode, the CountdownTimer can be
tested so a new test has been added to unit test the functions
of the CountdownTimer component.