feat(Data): rename InternalSetting to Restricted and add more options
The InternalSetting attribute has now been renamed to Restricted and
comes with some additional options that can be applied to a field.

It's now possible to mark a restricted field as ReadOnly (either always
or only when the application is playing and the component is disabled)
and it's possible to still colour the field grey to create a muted
look.

Any existing [InternalSetting] attribute can be directly replaced with
[Restricted] as the default options should just work in a direct
replacement.